Move TestDockerWorkerSchemaMatchesD2G to a posix-only file
The multiuser Windows build embeds schemas/multiuser_windows.yml, which
has no "Docker Worker payload" branch at all (d2g doesn't support
Windows), so the test had nothing to compare there and failed CI.

darwin/linux/freebsd all embed the same schemas/*_posix.yml source, so
the test still has real content to check on those platforms even
though d2g only executes on Linux at runtime.
